tests/requirements.txt
attrs==20.3.0
bcrypt==3.2.0
cassandra-driver==3.25.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
distro==1.5.0
docker==4.4.4
docker-compose==1.28.6
dockerpty==0.4.1
docopt==0.6.2
geomet==0.2.1.post1
idna==2.10
iniconfig==1.1.1
jsonschema==3.2.0
kafka-python==2.0.2
packaging==20.9
paramiko==2.7.2
pluggy==0.13.1
py==1.10.0
pycparser==2.20
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.2
pytest-docker==0.10.1
python-dotenv==0.16.0
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
six==1.15.0
texttable==1.6.3
toml==0.10.2
urllib3==1.26.4
websocket-client==0.58.0
